+++ src/main/java/nks/contacts/domain/pagination/PageableSource.java
@@ -3,6 +3,6 @@ package nks.contacts.domain.pagination;
 import java.util.List;
 
 public interface PageableSource<T> {
-	Integer itemCount();
+	Long itemCount();
     List<T> pagedItems(Integer offset, Integer count);
 }
